1.  重构文件夹 2. 使用npm做包管理 3. 使用gulp做自动化打包部署
