+++ example/001gettingstart/index.html
@@ -8,10 +8,10 @@
 	<meta name="viewport" content="initial-scale=1, maximum-scale=1, user-scalable=no, width=device-width, minimal-ui">
 	<!-- iphone设备中的safari私有meta标签，它表示：允许全屏模式浏览 -->
 	<meta name="apple-mobile-web-app-capable" content="yes">
-	<link rel="stylesheet" href="example/css/style.css">
+	<link rel="stylesheet" href="css/style.css">
 </head>
 <body>
-	<div style="width:320px;margin: 0 auto;">
+	<div style="width:100%;margin: 0 auto;">
 		<div class="publish-article-title">
 			<div class="title-tips">标题</div>
 			<input type="text" id="title" class="w100" placeholder="文章标题">
@@ -29,8 +29,8 @@
 			</div>
 		</div>
 	</div>
-	<script src="example/js/jquery-1.11.2.js"></script>
-	<script src="artEditor.min.js"></script>
-	<script src="example/js/index.js"></script>
+	<script src="js/jquery-1.11.2.js"></script>
+	<script src="../../dist/index.min.js"></script>
+	<script src="js/index.js"></script>
 </body>
 </html>
