--- example/js/index.js
@@ -1,23 +1,0 @@
-$(function() {
-	"use strict";
-
-	$('#content').artEditor({
-		imgTar: '#imageUpload',
-		limitSize: 5,   // 兆
-		showServer: false,
-		uploadUrl: '',
-		data: {},
-		uploadField: 'image',
-		placeholader: '<p>请输入文章正文内容</p>',
-		validHtml: ["<br/>"],
-		formInputId: 'target',
-		uploadSuccess: function(res) {
-			// return img url
-			return res.path;
-		},
-		uploadError: function(res) {
-			// something error
-			console.log(res);
-		}
-	});
-});
